backend/requirements.txt
# FastAPI Backend for Physical AI Textbook
# Authentication & Rate Limiting API

# Web Framework
fastapi>=0.109.2
uvicorn[standard]>=0.27.1

# JWT & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.35

# RAG - Vector Database & Embeddings
chromadb>=0.4.22
google-generativeai>=0.3.2
sentence-transformers>=2.2.2

# Validation
pydantic>=2.6.1
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.23.4
httpx>=0.26.0

# Development
python-dotenv>=1.0.1
